GH-34 - moved imageupload to separate component
